fix(api): update z.number() to z.number().int() for integer validation (#1713)

+++ apps/api/src/routes/v1_keys_verifyKey.ts
@@ -94,25 +94,28 @@ A key could be invalid for a number of reasons, for example if it has expired, h
                     stripeCustomerId: "cus_1234",
                   },
                 }),
-              expires: z.number().optional().openapi({
+              expires: z.number().int().optional().openapi({
                 description:
                   "The unix timestamp in milliseconds when the key will expire. If this field is null or undefined, the key is not expiring.",
                 example: 123,
               }),
               ratelimit: z
                 .object({
-                  limit: z.number().openapi({
+                  limit: z.number().int().openapi({
                     description: "Maximum number of requests that can be made inside a window",
                     example: 10,
                   }),
-                  remaining: z.number().openapi({
+                  remaining: z.number().int().openapi({
                     description: "Remaining requests after this verification",
                     example: 9,
                   }),
-                  reset: z.number().openapi({
-                    description: "Unix timestamp in milliseconds when the ratelimit will reset",
-                    example: Date.now() + 1000 * 60 * 60,
-                  }),
+                  reset: z
+                    .number()
+                    .int()
+                    .openapi({
+                      description: "Unix timestamp in milliseconds when the ratelimit will reset",
+                      example: Date.now() + 1000 * 60 * 60,
+                    }),
                 })
                 .optional()
                 .openapi({
@@ -124,7 +127,7 @@ A key could be invalid for a number of reasons, for example if it has expired, h
                     reset: Date.now() + 1000 * 60 * 60,
                   },
                 }),
-              remaining: z.number().optional().openapi({
+              remaining: z.number().int().optional().openapi({
                 description:
                   "The number of requests that can be made with this key before it becomes invalid. If this field is null or undefined, the key has no request limit.",
                 example: 1000,
